Refactor handler and filter handling
